Move peerStreamPool into its own file, and don't push stream errors onto waiters unless necessary
