+++ go.mod
@@ -2,6 +2,7 @@ module github.com/libp2p/go-libp2p-kad-dht
 
 require (
 	github.com/gogo/protobuf v1.2.1
+	github.com/golang/mock v1.2.0 // indirect
 	github.com/golang/protobuf v1.3.1 // indirect
 	github.com/hashicorp/golang-lru v0.5.1
 	github.com/ipfs/go-cid v0.0.1
